# gateway-runtime/docs/swagger.yaml
definitions:
...
  auth.AuthRequest:
    properties:
      password:
        example: rahasia1234
        minLength: 6
        type: string
      username:
        example: auditor_senior
        type: string
    required:
    - password
    - username
    type: object
  auth.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  auth.LoginResponse:
    properties:
      message:
        type: string
      token:
        type: string
    type: object
  auth.ProfileResponseData:
    properties:
      client_id:
        type: string
      company_name:
        type: string
      created_at:
        type: string
      full_name:
        type: string
      id:
        type: string
      role:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  auth.ProfileUpdateRequest:
    properties:
      current_password:
        type: string
      full_name:
        type: string
      new_password:
        type: string
      username:
        minLength: 4
        type: string
    required:
    - username
    type: object
  auth.RegisterRequest:
    properties:
      client_id:
        example: a1b2c3d4-e5f6-7890-1234-56789abcdef0
        type: string
      password:
        example: rahasia1234
        minLength: 6
        type: string
      username:
        example: auditor_senior
        type: string
    required:
    - client_id
    - password
    - username
    type: object
  auth.RegisterResponse:
    properties:
      message:
        type: string
      user:
        $ref: '#/definitions/auth.UserResponseData'
    type: object
  auth.UpdateProfileResponse:
    properties:
      message:
        type: string
      token:
        type: string
      user:
        $ref: '#/definitions/auth.ProfileResponseData'
    type: object
  auth.UserResponseData:
    properties:
      client_id:
        type: string
      id:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
...
info:
  contact:
    email: support@auditchain.local
    name: API Support
  description: API Enterprise untuk sistem audit log berbasis Blockchain dan Merkle
    Tree.
  termsOfService: http://swagger.io/terms/
  title: AuditChain Gateway API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Autentikasi pengguna dan mendapatkan token JWT untuk akses API.
      parameters:
      - description: Kredensial Login
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login berhasil dan mengembalikan token
          schema:
            $ref: '#/definitions/auth.LoginResponse'
        "400":
          description: Format request tidak valid
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: Username atau Password salah!
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "403":
          description: Akun dinonaktifkan karena client tidak aktif
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: Gagal mencetak token keamanan
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: User login
      tags:
      - Auth
  /auth/me:
    get:
      consumes:
      - application/json
      description: Mendapatkan informasi profil dari pengguna yang sedang login (berdasarkan
        JWT).
      produces:
      - application/json
      responses:
        "200":
          description: Data profil user
          schema:
            $ref: '#/definitions/auth.ProfileResponseData'
        "401":
          description: Token tidak memiliki identitas user yang valid
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "404":
          description: Profil user tidak ditemukan
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - Auth
    put:
      consumes:
      - application/json
      description: Memperbarui informasi profil pengguna yang sedang login. Bisa juga
        untuk mengubah password.
      parameters:
      - description: Data Profil Baru
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.ProfileUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Profil berhasil diperbarui dan mengembalikan token baru jika
            password diubah
          schema:
            $ref: '#/definitions/auth.UpdateProfileResponse'
        "400":
          description: Validasi input gagal (misal username terlalu pendek)
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: Token tidak valid atau Password saat ini tidak sesuai
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "409":
          description: Username sudah digunakan
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: Gagal memperbarui profil
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Mendaftarkan pengguna baru (auditor/admin) ke dalam perusahaan
        (client) tertentu.
      parameters:
      - description: Data Pendaftaran
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Pengguna berhasil didaftarkan
          schema:
            $ref: '#/definitions/auth.RegisterResponse'
        "400":
          description: Format tidak valid atau client_id belum diisi
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "404":
          description: Perusahaan (Client ID) tidak terdaftar di sistem
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "409":
          description: Username sudah digunakan
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: Gagal memproses pendaftaran
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Register a new user
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token dengan format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"